+++ src/components/movie/Detail.vue
