+++ src/pages/ActivityPage.vue
